refactor: move helper to chain utils
