+++ src/utils/__tests__/chains.test.ts
@@ -1,4 +1,4 @@
-import { hasFeature, getBlockExplorerLink, FEATURES } from '@/utils/chains'
+import { hasFeature, getBlockExplorerLink, FEATURES, getLatestSafeVersion } from '@/utils/chains'
 import { CONFIG_SERVICE_CHAINS } from '@/tests/mocks/chains'
 
 describe('chains', () => {
@@ -34,6 +34,20 @@ describe('chains', () => {
       ).toEqual({
         href: 'https://etherscan.io/tx/0x123436456456754735474574575475675435353453465645645656',
         title: 'View on etherscan.io',
+      })
+    })
+  })
+
+  describe('chains', () => {
+    describe('getLatestSafeVersion', () => {
+      it('should return 1.4.1 on supported networks', () => {
+        expect(getLatestSafeVersion('1')).toEqual('1.4.1')
+        expect(getLatestSafeVersion('137')).toEqual('1.4.1')
+        expect(getLatestSafeVersion('11155111')).toEqual('1.4.1')
+      })
+
+      it('should return 1.3.0 on networks where 1.4.1 is not released', () => {
+        expect(getLatestSafeVersion('324')).toEqual('1.3.0')
       })
     })
   })
