+++ src/utils/chains.ts
@@ -1,6 +1,13 @@
 import { AppRoutes } from '@/config/routes'
 import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import { getExplorerLink } from './gateway'
+import { LATEST_SAFE_VERSION } from '@/config/constants'
+import { type SafeVersion } from '@safe-global/safe-core-sdk-types'
+import { getSafeSingletonDeployment } from '@safe-global/safe-deployments'
+import semverSatisfies from 'semver/functions/satisfies'
+
+/** This version is used if a network does not have the LATEST_SAFE_VERSION deployed yet */
+const FALLBACK_SAFE_VERSION = '1.3.0' as const
 
 export enum FEATURES {
   ERC721 = 'ERC721',
@@ -52,4 +59,17 @@ export const isRouteEnabled = (route: string, chain?: ChainInfo) => {
   if (!chain) return false
   const featureRoute = FeatureRoutes[route]
   return !featureRoute || hasFeature(chain, featureRoute)
+}
+
+export const getLatestSafeVersion = (chainId?: string): SafeVersion => {
+  // Without version filter it will always return the LATEST_SAFE_VERSION constant to avoid automatically updating to the newest version if the deployments change
+  const latestVersion = (getSafeSingletonDeployment({ network: chainId, released: true })?.version ??
+    FALLBACK_SAFE_VERSION) as SafeVersion
+
+  // The version needs to be smaller or equal to the
+  if (semverSatisfies(latestVersion, `<=${LATEST_SAFE_VERSION}`)) {
+    return latestVersion
+  } else {
+    return LATEST_SAFE_VERSION as SafeVersion
+  }
 }
